+++ core/src/main/kotlin/ca/allanwang/kau/swipe/ViewDragHelper.java
@@ -25,7 +25,7 @@ import static ca.allanwang.kau.swipe.SwipeBackHelperKt.SWIPE_EDGE_TOP;
  * of useful operations and state tracking for allowing a user to drag and reposition
  * views within their parent ViewGroup.
  * <p>
- * This is an extension of {@link androidx.core.widget.ViewDragHelper}
+ * This is an extension of {@link androidx.customview.widget.ViewDragHelper}
  * Along with additional methods defined in {@link ViewDragHelperExtras}
  */
 class ViewDragHelper implements ViewDragHelperExtras {